Refactor styles and routes outside of main
